Improved the travis build script a bit.

+++ tests/bin/setup.mysql.sh
@@ -3,7 +3,17 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )";
 . $DIR/setup.base.sh;
 
-mysql -u$DB_USER -e '\
+mysql=`which mysql`;
+if [ "$mysql" = "" ]; then
+    mysql=`which mysql5`;
+fi
+
+if [ "$mysql" = "" ]; then
+    echo "Can not find mysql binary. Is it installed?";
+    exit;
+fi
+
+$mysql -u$DB_USER -e '\
 SET FOREIGN_KEY_CHECKS = 0; \
 DROP DATABASE IF EXISTS test; \
 DROP SCHEMA IF EXISTS second_hand_books; \
@@ -11,14 +21,18 @@ DROP SCHEMA IF EXISTS contest; \
 DROP SCHEMA IF EXISTS bookstore_schemas; \
 SET FOREIGN_KEY_CHECKS = 1; \
 ';
-echo "Delete databases and schemas in MySQL database.";
+check;
 
-mysql -u$DB_USER -e '\
+$mysql -u$DB_USER -e '\
 CREATE DATABASE test; \
 CREATE SCHEMA bookstore_schemas; \
 CREATE SCHEMA contest; \
 CREATE SCHEMA second_hand_books; \
 ';
-echo "MySQL with the databases and schemas of the testsuite equipped.";
+check;
 
-php $DIR/../../bin/propel test:prepare --vendor="$DB" --dsn="$DB:dbname=test" --user="$DB_USER"
+if [ "$DB_PW" = "" ]; then
+    php $DIR/../../bin/propel test:prepare --vendor="$DB" --dsn="$DB:dbname=test" --user="$DB_USER";
+else
+    php $DIR/../../bin/propel test:prepare --vendor="$DB" --dsn="$DB:dbname=test" --user="$DB_USER" --password="$DB_PW";
+fi
